requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.30.0
tokenizers==0.13.0

numpy
Pillow
tqdm
requests
matplotlib
scikit-learn

opencv-python

# optional but recommended
sentencepiece
einops
